patch: Slot packing

+++ contracts/oracle/utils/SequencerCheck.sol
@@ -6,7 +6,7 @@ import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 contract SequencerCheck {
     // sequencer uptime feed
     AggregatorV3Interface public immutable SEQUENCER_UPTIME_FEED; // If not set, assume it's on a L1
-    uint256 public GRACE_PERIOD_TIME = 3600; // 1 hour
+    uint32 public GRACE_PERIOD_TIME = 3600; // 1 hour
 
     bytes32 public constant SEQUENCER_ROLE = keccak256("SEQUENCER_ROLE");
